controller + some refactoring
