+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/DummyClientImpl.java
@@ -1,15 +1,15 @@
 package de.muenchen.oss.wahllokalsystem.adminservice.client;
 
 import de.muenchen.oss.wahllokalsystem.adminservice.configuration.Profiles;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.AWerteClient;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.KonfigurierterWahltagClient;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.KonfigurierterWahltagModel;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahlbezirkArtModel;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahlbezirkModel;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahlbezirkeClient;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahltagModel;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahltageClient;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahltermindatenClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.AWerteClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.KonfigurierterWahltagClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.KonfigurierterWahltagModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahlbezirkArtModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahlbezirkModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahlbezirkeClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahltagModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahltageClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahltermindatenClient;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.WlsException;
 import java.time.LocalDate;
 import java.util.List;
