+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/basisdaten/WahlbezirkeClientImpl.java
@@ -3,8 +3,8 @@ package de.muenchen.oss.wahllokalsystem.adminservice.client.basisdaten;
 import de.muenchen.oss.wahllokalsystem.adminservice.configuration.Profiles;
 import de.muenchen.oss.wahllokalsystem.adminservice.eai.basisdaten.client.WahlbezirkeControllerApi;
 import de.muenchen.oss.wahllokalsystem.adminservice.exception.ExceptionConstants;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahlbezirkModel;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahlbezirkeClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahlbezirkModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahlbezirkeClient;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.WlsException;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import java.util.List;
