+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/basisdaten/WahlbezirkeClientMapper.java
@@ -1,7 +1,7 @@
 package de.muenchen.oss.wahllokalsystem.adminservice.client.basisdaten;
 
 import de.muenchen.oss.wahllokalsystem.adminservice.eai.basisdaten.model.WahlbezirkDTO;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahlbezirkModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahlbezirkModel;
 import java.util.List;
 import org.mapstruct.Mapper;
