+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/basisdaten/WahltagClientMapper.java
@@ -1,7 +1,7 @@
 package de.muenchen.oss.wahllokalsystem.adminservice.client.basisdaten;
 
 import de.muenchen.oss.wahllokalsystem.adminservice.eai.basisdaten.model.WahltagDTO;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahltagModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahltagModel;
 import java.util.List;
 import org.mapstruct.Mapper;
