+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/basisdaten/WahltageClientImpl.java
@@ -3,8 +3,8 @@ package de.muenchen.oss.wahllokalsystem.adminservice.client.basisdaten;
 import de.muenchen.oss.wahllokalsystem.adminservice.configuration.Profiles;
 import de.muenchen.oss.wahllokalsystem.adminservice.eai.basisdaten.client.WahltageControllerApi;
 import de.muenchen.oss.wahllokalsystem.adminservice.exception.ExceptionConstants;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahltagModel;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.WahltageClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahltagModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.WahltageClient;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.WlsException;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import java.util.List;
