+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/infomanagement/KonfigurierterWahltagClientImpl.java
@@ -3,8 +3,8 @@ package de.muenchen.oss.wahllokalsystem.adminservice.client.infomanagement;
 import de.muenchen.oss.wahllokalsystem.adminservice.configuration.Profiles;
 import de.muenchen.oss.wahllokalsystem.adminservice.eai.infomanagement.client.KonfigurierterWahltagControllerApi;
 import de.muenchen.oss.wahllokalsystem.adminservice.exception.ExceptionConstants;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.KonfigurierterWahltagClient;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.KonfigurierterWahltagModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.KonfigurierterWahltagClient;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.KonfigurierterWahltagModel;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.WlsException;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import lombok.RequiredArgsConstructor;
