+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/client/infomanagement/KonfigurierterWahltagClientMapper.java
@@ -1,7 +1,10 @@
 package de.muenchen.oss.wahllokalsystem.adminservice.client.infomanagement;
 
+import static de.muenchen.oss.wahllokalsystem.adminservice.eai.infomanagement.model.KonfigurierterWahltagDTO.WahltagStatusEnum.AKTIV;
+import static de.muenchen.oss.wahllokalsystem.adminservice.eai.infomanagement.model.KonfigurierterWahltagDTO.WahltagStatusEnum.INAKTIV;
+
 import de.muenchen.oss.wahllokalsystem.adminservice.eai.infomanagement.model.KonfigurierterWahltagDTO;
-import de.muenchen.oss.wahllokalsystem.adminservice.service.KonfigurierterWahltagModel;
+import de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten.KonfigurierterWahltagModel;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
@@ -10,4 +13,8 @@ public interface KonfigurierterWahltagClientMapper {
 
     @Mapping(target = "wahltagStatus", source = "active")
     KonfigurierterWahltagDTO toClientDto(KonfigurierterWahltagModel model);
+
+    default KonfigurierterWahltagDTO.WahltagStatusEnum mapModelBooleanActiveToWahltagStatusEnumString(boolean active) {
+        return active ? AKTIV : INAKTIV;
+    }
 }
