--- wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/service/WahlbezirkArtModel.java
@@ -1,5 +1,0 @@
-package de.muenchen.oss.wahllokalsystem.adminservice.service;
-
-public enum WahlbezirkArtModel {
-    UWB, BWB
-}
