+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/service/wahltermindaten/KonfigurierterWahltagClient.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.adminservice.service;
+package de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten;
 
 public interface KonfigurierterWahltagClient {
