+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/service/wahltermindaten/WahlbezirkArtModel.java
@@ -1,0 +1,5 @@
+package de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten;
+
+public enum WahlbezirkArtModel {
+    UWB, BWB
+}
