+++ wls-admin-service/src/main/java/de/muenchen/oss/wahllokalsystem/adminservice/service/wahltermindaten/WahlbezirkeClient.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.adminservice.service;
+package de.muenchen.oss.wahllokalsystem.adminservice.service.wahltermindaten;
 
 import java.util.List;
